PG: Add metrics for wal files: count, size and age
PG10 introduces the `pg_ls_waldir` function which allows us to inspect
the WAL dir content without the need to run the agent locally.
Use it to export the count, total size and age of the WAL files.

Since there's no constraint on running `pg_ls_waldir` except to have
pg_monitor access, it would make sense to make it available by default
instead of opt-in as it was before.